requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
cssselect==1.2.0
dataclasses-json==0.5.7
dateparser==1.1.8
demjson3==3.0.6
facebook-scraper==0.2.59
facebook-sdk==3.1.0
facepy==1.0.12
fake-useragent==1.1.3
idna==3.4
importlib-metadata==6.6.0
lxml==4.9.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
parse==1.19.0
pychalk==2.0.1
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.5.5
requests==2.30.0
requests-html==0.10.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzlocal==5.0.1
urllib3==1.26.15
w3lib==2.1.1
websockets==10.4
zipp==3.15.0
